refactor: directory organisation for search related models
Relates #256
